Collect coroutine pipeline generics under stream
