+++ flash_stats.py
@@ -6,7 +6,7 @@ import os
 import numpy as np
 
 from lmatools.io.LMA_h5_file import read_flashes
-from lmatools.density_to_files import coroutine, Branchpoint
+from lmatools.stream.subset import coroutine, Branchpoint
 
 from lmatools.flashsort.flash_stats import hull_volume
 from six.moves import range
